+++ app/src/main/java/com/sopt/peekabookaos/data/repository/MyPageRepositoryImpl.kt
@@ -3,6 +3,8 @@ package com.sopt.peekabookaos.data.repository
 import com.sopt.peekabookaos.data.source.remote.MyPageDataSource
 import com.sopt.peekabookaos.domain.entity.User
 import com.sopt.peekabookaos.domain.repository.MyPageRepository
+import okhttp3.MultipartBody
+import okhttp3.RequestBody
 import javax.inject.Inject
 
 class MyPageRepositoryImpl @Inject constructor(
@@ -12,4 +14,13 @@ class MyPageRepositoryImpl @Inject constructor(
         kotlin.runCatching { myPageDataSource.getMyPage() }.map { response ->
             requireNotNull(response.data).toUser()
         }
+
+    override suspend fun patchProfileModify(
+        file: MultipartBody.Part?,
+        requestBodyMap: HashMap<String, RequestBody>
+    ): Result<Boolean> =
+        kotlin.runCatching { myPageDataSource.patchProfileModify(file, requestBodyMap) }
+            .map { response ->
+                response.success
+            }
 }
